@@ -31,9 +31,8 @@
         }
     ],
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "ext-dom": "*",
-        "ext-json": "*",
         "ext-mbstring": "*",
         "alexkart/curl-builder": "^1.0",
         "cebe/markdown": "^1.2",
@@ -48,16 +47,16 @@
         "yiisoft/injector": "^1.0"
     },
     "require-dev": {
-        "httpsoft/http-message": "^1.0.9",
-        "maglnet/composer-require-checker": "^4.4",
-        "phpunit/phpunit": "^9.5",
+        "httpsoft/http-message": "^1.1.6",
+        "maglnet/composer-require-checker": "^4.7.1",
+        "phpunit/phpunit": "^10.5.44",
         "psr/event-dispatcher": "^1.0",
-        "rector/rector": "^2.0",
-        "roave/infection-static-analysis-plugin": "^1.16",
-        "spatie/phpunit-watcher": "^1.23",
-        "vimeo/psalm": "^4.30|^5.25",
-        "yiisoft/di": "^1.1",
-        "yiisoft/test-support": "^3.0"
+        "rector/rector": "^2.0.7",
+        "roave/infection-static-analysis-plugin": "^1.35",
+        "spatie/phpunit-watcher": "^1.24",
+        "vimeo/psalm": "^5.26.1|^6",
+        "yiisoft/di": "^1.3",
+        "yiisoft/test-support": "^3.0.1"
     },
     "autoload": {
         "psr-4": {
@@ -79,6 +78,7 @@
     },
     "config": {
         "sort-packages": true,
+        "bump-after-update": "dev",
         "allow-plugins": {
             "infection/extension-installer": true,
             "composer/package-versions-deprecated": true
