@@ -11,8 +11,9 @@ use Throwable;
  */
 final class ApplicationError
 {
-    public function __construct(private Throwable $throwable)
-    {
+    public function __construct(
+        private readonly Throwable $throwable,
+    ) {
     }
 
     public function getThrowable(): Throwable
